Module Storage.Contract

Storage from this submodule must only be accessed through the module `Contract`.

module Global_counter : sig ... end

val fold : Raw_context.t -> init:'a -> f:(Contract_repr.t -> 'a -> 'a Tezos_protocol_environment.Lwt.t) -> 'a Tezos_protocol_environment.Lwt.t

The domain of alive contracts

val list : Raw_context.t -> Contract_repr.t list Tezos_protocol_environment.Lwt.t

module Balance : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = Tez_repr.t and type t := Raw_context.t

All the tez possessed by a contract, including rolls and change

module Frozen_deposits : Storage_sigs.Indexed_data_storage with type key = Cycle_repr.t and type value = Tez_repr.t and type t = Raw_context.t * Contract_repr.t

Frozen balance, see 'delegate_storage.mli' for more explanation. Always update `Delegates_with_frozen_balance` accordingly.

module Frozen_fees : Storage_sigs.Indexed_data_storage with type key = Cycle_repr.t and type value = Tez_repr.t and type t = Raw_context.t * Contract_repr.t

module Frozen_rewards : Storage_sigs.Indexed_data_storage with type key = Cycle_repr.t and type value = Tez_repr.t and type t = Raw_context.t * Contract_repr.t

module Manager : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = Manager_repr.t and type t := Raw_context.t

The manager of a contract

module Delegate : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = Tezos_protocol_environment.Signature.Public_key_hash.t and type t := Raw_context.t

The delegate of a contract, if any.

module Delegated : Storage_sigs.Data_set_storage with type elt = Contract_repr.t and type t = Raw_context.t * Contract_repr.t

All contracts (implicit and originated) that are delegated, if any

module Inactive_delegate : Storage_sigs.Data_set_storage with type elt = Contract_repr.t and type t = Raw_context.t

module Delegate_desactivation : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = Cycle_repr.t and type t := Raw_context.t

The cycle where the delegate should be deactivated.

module Counter : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = Tezos_protocol_environment.Z.t and type t := Raw_context.t

module Code : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Contract_repr.t and type value = Script_repr.lazy_expr and type t := Raw_context.t

module Storage : Storage_sigs.Non_iterable_indexed_carbonated_data_storage with type key = Contract_repr.t and type value = Script_repr.lazy_expr and type t := Raw_context.t

module Used_storage_space : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = Tezos_protocol_environment.Z.t and type t := Raw_context.t

Current storage space in bytes. Includes code, global storage and big map elements.

module Paid_storage_space : Storage_sigs.Indexed_data_storage with type key = Contract_repr.t and type value = Tezos_protocol_environment.Z.t and type t := Raw_context.t

Maximal space available without needing to burn new fees.