Module Make.Tez_repr

type t

type tez = t

include Qty_repr.S with type qty := t

type qty

type Tezos_protocol_environment.Error_monad.error += Addition_overflow of qty * qty | Subtraction_underflow of qty * qty | Multiplication_overflow of qty * int64 | Negative_multiplicator of qty * int64 | Invalid_divisor of qty * int64

val id : string

val zero : qty

val one_mutez : qty

val one_cent : qty

val fifty_cents : qty

val one : qty

val (-?) : qty -> qty -> qty Tezos_protocol_environment.Error_monad.tzresult

val (+?) : qty -> qty -> qty Tezos_protocol_environment.Error_monad.tzresult

val (*?) : qty -> int64 -> qty Tezos_protocol_environment.Error_monad.tzresult

val (/?) : qty -> int64 -> qty Tezos_protocol_environment.Error_monad.tzresult

val to_mutez : qty -> int64

val of_mutez : int64 -> qty option

of_mutez n (micro tez) is None if n is negative

val of_mutez_exn : int64 -> qty

of_mutez_exn n fails if n is negative. It should only be used at toplevel for constants.

val add_exn : qty -> qty -> qty

It should only be used at toplevel for constants.

val mul_exn : qty -> int -> qty

It should only be used at toplevel for constants.

val encoding : qty Tezos_protocol_environment.Data_encoding.t

val to_int64 : qty -> int64

include Tezos_protocol_environment.Compare.S with type t := qty

type t

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (>) : t -> t -> bool

val compare : t -> t -> int

val equal : t -> t -> bool

val max : t -> t -> t

val min : t -> t -> t

val pp : Tezos_protocol_environment.Format.formatter -> qty -> unit

val of_string : string -> qty option

val to_string : qty -> string